Agentic_AI_with_RAG_and_Memory/requirements.txt
phidata
python-dotenv
packaging
fastapi
uvicorn
groq
sqlalchemy
pgvector
psycopg[binary]
pypdf
typer
openai
sentence-transformers